repositoryController tests split per methods
