+++ backend/tests/Designer.Tests/Controllers/ResourceAdminControllerTests.cs
@@ -12,12 +12,12 @@ using Xunit;
 
 namespace Designer.Tests.Controllers
 {
-    public class ResourceAdminControllerTests : ApiTestsBase<RepositoryController, RepositoryControllerTests>
+    public class ResourceAdminControllerTests : ApiTestsBase<ResourceAdminController, ResourceAdminControllerTests>
     {
         private readonly string _versionPrefix = "/designer/api";
         private readonly Mock<IGitea> _giteaMock;
 
-        public ResourceAdminControllerTests(WebApplicationFactory<RepositoryController> factory) : base(factory)
+        public ResourceAdminControllerTests(WebApplicationFactory<ResourceAdminController> factory) : base(factory)
         {
             _giteaMock = new Mock<IGitea>();
         }
